+++ examples/annulus.jl
@@ -16,8 +16,8 @@ sequence = mr.perfect_dwi(TR=TR, TE=TE, bval=bval)
 
 snap = mr.Snapshot(3000);
 
-simulation = mr.Simulation([sequence], diffusivity=3., geometry=geometry, timestep=0.1);
+simulation = mr.Simulation([sequence], diffusivity=3., geometry=geometry);
 
 mr.evolve(snap, simulation, 2.);
-@time mr.evolve(snap, simulation, 200.);
-@profview mr.evolve(snap, simulation, 200.);
+@time mr.evolve(snap, simulation, 20.);
+@profview mr.evolve(snap, simulation, 20.);