+++ src/geometry/base/random_draws.jl
@@ -1,12 +1,16 @@
-function draw_radii(box_size::SVector{N, <:Real}, target_density::Real, distribution::Distributions.ContinuousUnivariateDistribution) where {N}
+function draw_radii(box_size::SVector{N, <:Real}, target_density::Real, distribution::Distributions.ContinuousUnivariateDistribution, allowed_range::Tuple{<:Real, <:Real}) where {N}
     if target_density > 1 || target_density < 0
         error("Target density should be a number between 0 and 1")
     end
+    @assert allowed_range[2] > allowed_range[1]
     volume = 0.
     target_volume = prod(box_size) * target_density
     samples = Float[]
     while volume < target_volume
         new_value = rand(distribution)
+        if new_value < allowed_range[1] || new_value > allowed_range[2]
+            continue 
+        end
         push!(samples, new_value)
         if N == 3
             volume += 4//3 * π * new_value^3
@@ -54,7 +58,7 @@ end
 
 
 """
-    random_positions_radii(box_size, target_density, n_dimensions; distribution=Gamma, mean=1., variance=1., max_iter=1000)
+    random_positions_radii(box_size, target_density, n_dimensions; distribution=Gamma, mean=1., variance=1., max_iter=1000, min_radius=0.1, max_radius=Inf)
 
 Randomly distributes circles or spheres in space.
 
@@ -66,10 +70,13 @@ Arguments:
 - `mean`: mean of the gamma distribution (ignored if `distribution` explicitly set)
 - `variance`: variance of the gamma distribution (ignored if `distribution` explicitly set)
 - `max_iter`: maximum number of iterations to try to prevent the circles/spheres from overlapping. An error is raised if they still overlap after this number of iterations.
+- `min_radius`: samples from the distribution below this radius will be rejected (in um).
+- `max_radius`: samples from the distribution above this radius will be rejected (in um).
 """
 function random_positions_radii(
     box_size, target_density::Real, ndim::Int;
-    distribution=nothing, mean=1., variance=1., max_iter=1000
+    distribution=nothing, mean=1., variance=1., max_iter=1000,
+    min_radius=0.1, max_radius=Inf,
 )
     if isnothing(distribution)
         a = mean * mean / variance
@@ -82,7 +89,7 @@ function random_positions_radii(
         box_size = SVector{ndim, Float}(box_size)
     end
     target_density = Float(target_density)
-    radii = draw_radii(box_size, target_density, distribution)
+    radii = draw_radii(box_size, target_density, distribution, (min_radius, max_radius))
     start_positions = [SVector{ndim, Float}((rand(ndim) .- 0.5) .* box_size) for _ in 1:length(radii)]
     return eliminate_overlap(start_positions, radii, box_size; max_iter=max_iter), radii
 end