+++ src/geometry/geometry_struct.jl
@@ -76,8 +76,9 @@ end
 Computes the MRI parameters for the spin based on some global settings (`global_props`) and any overriding of those settings in the `geometry`.
 """
 function inside_MRI_properties(geom::Geometry, position::PosVector, global_props::MRIProperties)
-    return merge_mri_parameters([inside_MRI_properties(o, position) for o in geom.obstructions], global_props)
+    return merge_mri_parameters((inside_MRI_properties(o, position) for o in geom.obstructions), global_props)
 end
+inside_MRI_properties(geom::Geometry{0}, position::PosVector, global_props::MRIProperties) = global_props
 
 inside_MRI_properties(geom::Geometry, spin::Spin, global_props::MRIProperties) = inside_MRI_properties(geom, spin.position, global_props)
 inside_MRI_properties(geom::Geometry, position::AbstractVector, global_props::MRIProperties) = inside_MRI_properties(geom, PosVector(position), global_props)